requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
feedparser>=6.0.0
aiohttp==3.9.1
atproto==0.0.31
requests>=2.26.0
PyNaCl==1.5.0  # For voice support
lxml==4.9.3  # For better XML parsing
html5lib==1.1  # For better HTML parsing
instaloader>=4.10.0
beautifulsoup4>=4.9.3